feat: Add 'AS3MultinameUpgrader' type for upgrading the package, and class names of AS3 objects
The upgrade process works by comparing the namespace of a trait, against the qualified name and namespace of its' container(class/instance). If the namespace of the trait does not match, the upgrader assumes that the trait carries the real package name, as well as the qualified name of the container it resides in.